Adding updated example and fix to CI.
