+++ mne_bids/path.py
@@ -574,7 +574,7 @@ class BIDSPath(object):
         self._check()
         return self
 
-    def match(self):
+    def match(self, return_json=False):
         """Get a list of all matching paths in the root directory.
 
         Performs a recursive search, starting in ``.root`` (if set), based on
@@ -601,6 +601,9 @@ class BIDSPath(object):
         fnames = self.root.rglob(search_str)
         # Only keep files (not directories), and omit the JSON sidecars.
         fnames = [f.name for f in fnames if f.is_file()]
+        if not return_json:
+            fnames = [f for f in fnames if not f.endswith('.json')]
+
         fnames = _filter_fnames(fnames, suffix=self.suffix,
                                 extension=self.extension,
                                 **self.entities)
